requirements.txt
Pillow
django-compressor
